# Cat-Companion/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cat Care API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Cat {
  id: number;
  name: string;
  breed?: string | null;
  birthdate?: string | null;
  weightKg?: number | null;
  photoUrl?: string | null;
  notes?: string | null;
  favouriteToy?: string | null;
  favouriteFood?: string | null;
  personalityTraits?: string | null;
  favouriteSpot?: string | null;
  favouriteSound?: string | null;
  favouriteHuman?: string | null;
  traitScores?: string | null;
  awards?: string | null;
  createdAt: string;
}

export interface CreateCatBody {
  name: string;
  breed?: string | null;
  birthdate?: string | null;
  weightKg?: number | null;
  photoUrl?: string | null;
  notes?: string | null;
  favouriteToy?: string | null;
  favouriteFood?: string | null;
  personalityTraits?: string | null;
  favouriteSpot?: string | null;
  favouriteSound?: string | null;
  favouriteHuman?: string | null;
  traitScores?: string | null;
  awards?: string | null;
}

export interface UpdateCatBody {
  name?: string;
  breed?: string | null;
  birthdate?: string | null;
  weightKg?: number | null;
  photoUrl?: string | null;
  notes?: string | null;
  favouriteToy?: string | null;
  favouriteFood?: string | null;
  personalityTraits?: string | null;
  favouriteSpot?: string | null;
  favouriteSound?: string | null;
  favouriteHuman?: string | null;
  traitScores?: string | null;
  awards?: string | null;
}

export interface Feeding {
  id: number;
  catId: number;
  fedAt: string;
  foodType?: string | null;
  amountGrams?: number | null;
  notes?: string | null;
}

export interface LogFeedingBody {
  fedAt: string;
  foodType?: string | null;
  amountGrams?: number | null;
  notes?: string | null;
}

export interface VetAppointment {
  id: number;
  catId: number;
  date: string;
  type: string;
  vetName?: string | null;
  notes?: string | null;
  completed: boolean;
}

export interface CreateVetAppointmentBody {
  date: string;
  type: string;
  vetName?: string | null;
  notes?: string | null;
  completed?: boolean;
}

export interface UpdateVetAppointmentBody {
  date?: string;
  type?: string;
  vetName?: string | null;
  notes?: string | null;
  completed?: boolean;
}

export interface WeightLog {
  id: number;
  catId: number;
  weightKg: number;
  loggedAt: string;
  notes?: string | null;
}

export interface LogWeightBody {
  weightKg: number;
  loggedAt: string;
  notes?: string | null;
}

export interface DashboardSummary {
  totalCats: number;
  feedingsToday: number;
  upcomingAppointments: number;
  overdueAppointments: number;
}

export interface AppointmentWithCat {
  id: number;
  catId: number;
  catName: string;
  catPhotoUrl?: string | null;
  date: string;
  type: string;
  vetName?: string | null;
  notes?: string | null;
  completed: boolean;
}

export interface FeedingWithCat {
  id: number;
  catId: number;
  catName: string;
  catPhotoUrl?: string | null;
  fedAt: string;
  foodType?: string | null;
  amountGrams?: number | null;
  notes?: string | null;
}

export interface Milestone {
  id: number;
  catId: number;
  title: string;
  ageDescription?: string | null;
  photoUrl?: string | null;
  notes?: string | null;
  capturedAt: string;
}

export interface CreateMilestoneBody {
  title: string;
  ageDescription?: string | null;
  photoUrl?: string | null;
  notes?: string | null;
  capturedAt: string;
}

export interface UpdateMilestoneBody {
  title?: string;
  ageDescription?: string | null;
  photoUrl?: string | null;
  notes?: string | null;
  capturedAt?: string;
}

export interface MilestoneWithCat {
  id: number;
  catId: number;
  catName: string;
  catPhotoUrl?: string | null;
  title: string;
  ageDescription?: string | null;
  photoUrl?: string | null;
  notes?: string | null;
  capturedAt: string;
}

export interface AlbumEntry {
  id: number;
  catId: number;
  photoUrl?: string | null;
  caption?: string | null;
  capturedAt: string;
}

export interface CreateAlbumEntryBody {
  photoUrl?: string | null;
  caption?: string | null;
  capturedAt: string;
}

export interface UpdateAlbumEntryBody {
  photoUrl?: string | null;
  caption?: string | null;
  capturedAt?: string;
}

export interface Medication {
  id: number;
  catId: number;
  name: string;
  dose?: string | null;
  frequencyLabel?: string | null;
  startDate: string;
  endDate?: string | null;
  notes?: string | null;
  isActive: boolean;
  nextDoseAt?: string | null;
}

export interface CreateMedicationBody {
  name: string;
  dose?: string | null;
  frequencyLabel?: string | null;
  startDate: string;
  endDate?: string | null;
  notes?: string | null;
  isActive?: boolean;
  nextDoseAt?: string | null;
}

export interface UpdateMedicationBody {
  name?: string;
  dose?: string | null;
  frequencyLabel?: string | null;
  startDate?: string;
  endDate?: string | null;
  notes?: string | null;
  isActive?: boolean;
  nextDoseAt?: string | null;
}

export interface MarkMedicationGivenBody {
  nextDoseAt?: string | null;
}

export interface MedicationWithCat {
  id: number;
  catId: number;
  catName: string;
  catPhotoUrl?: string | null;
  name: string;
  dose?: string | null;
  frequencyLabel?: string | null;
  startDate: string;
  endDate?: string | null;
  notes?: string | null;
  isActive: boolean;
  nextDoseAt?: string | null;
}

export interface HealthRecord {
  id: number;
  catId: number;
  type: string;
  name: string;
  date: string;
  nextDueDate?: string | null;
  administeredBy?: string | null;
  batchNumber?: string | null;
  notes?: string | null;
}

export interface CreateHealthRecordBody {
  type: string;
  name: string;
  date: string;
  nextDueDate?: string | null;
  administeredBy?: string | null;
  batchNumber?: string | null;
  notes?: string | null;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface VetContact {
  id: number;
  catId?: number | null;
  clinicName: string;
  vetName?: string | null;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  isEmergency: boolean;
  notes?: string | null;
}

export interface CreateVetContactBody {
  catId?: number | null;
  clinicName: string;
  vetName?: string | null;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  isEmergency?: boolean;
  notes?: string | null;
}

export interface UpdateVetContactBody {
  catId?: number | null;
  clinicName?: string;
  vetName?: string | null;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  isEmergency?: boolean;
  notes?: string | null;
}

export type ListVetContactsParams = {
  catId?: number;
};