# Cat-Companion/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cat Care API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cats/{catId}/vet-appointments:
    get:
      operationId: listVetAppointments
      tags: [vet-appointments]
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of vet appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VetAppointment"
    post:
      operationId: createVetAppointment
      tags: [vet-appointments]
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVetAppointmentBody"
      responses:
        "201":
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VetAppointment"
  /vet-appointments/{id}:
    put:
      operationId: updateVetAppointment
      tags: [vet-appointments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVetAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VetAppointment"
    delete:
      operationId: deleteVetAppointment
      tags: [vet-appointments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /cats/{catId}/weight-logs:
    get:
      operationId: listWeightLogs
      tags: [weight-logs]
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of weight logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeightLog"
    post:
      operationId: logWeight
      tags: [weight-logs]
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogWeightBody"
      responses:
        "201":
          description: Logged weight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeightLog"
  /weight-logs/{id}:
    delete:
      operationId: deleteWeightLog
      tags: [weight-logs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /vet-contacts:
    get:
      operationId: listVetContacts
      tags: [vet-contacts]
      summary: List all vet contacts (optionally filter by catId)
      parameters:
        - name: catId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of vet contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VetContact"
    post:
      operationId: createVetContact
      tags: [vet-contacts]
      summary: Create a vet contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVetContactBody"
      responses:
        "201":
          description: Created vet contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VetContact"
  /vet-contacts/{id}:
    put:
      operationId: updateVetContact
      tags: [vet-contacts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVetContactBody"
      responses:
        "200":
          description: Updated vet contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VetContact"
    delete:
      operationId: deleteVetContact
      tags: [vet-contacts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    VetAppointment:
      type: object
      properties:
        id:
          type: integer
        catId:
          type: integer
        date:
          type: string
        type:
          type: string
        vetName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        completed:
          type: boolean
      required: [id, catId, date, type, completed]
    CreateVetAppointmentBody:
      type: object
      properties:
        date:
          type: string
        type:
          type: string
        vetName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        completed:
          type: boolean
      required: [date, type]
    UpdateVetAppointmentBody:
      type: object
      properties:
        date:
          type: string
        type:
          type: string
        vetName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        completed:
          type: boolean
    WeightLog:
      type: object
      properties:
        id:
          type: integer
        catId:
          type: integer
        weightKg:
          type: number
        loggedAt:
          type: string
        notes:
          type: string
          nullable: true
      required: [id, catId, weightKg, loggedAt]
    LogWeightBody:
      type: object
      properties:
        weightKg:
          type: number
        loggedAt:
          type: string
        notes:
          type: string
          nullable: true
      required: [weightKg, loggedAt]
...
    VetContact:
      type: object
      properties:
        id:
          type: integer
        catId:
          type: integer
          nullable: true
        clinicName:
          type: string
        vetName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        isEmergency:
          type: boolean
        notes:
          type: string
          nullable: true
      required: [id, clinicName, isEmergency]
    CreateVetContactBody:
      type: object
      properties:
        catId:
          type: integer
          nullable: true
        clinicName:
          type: string
        vetName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        isEmergency:
          type: boolean
        notes:
          type: string
          nullable: true
      required: [clinicName]
    UpdateVetContactBody:
      type: object
      properties:
        catId:
          type: integer
          nullable: true
        clinicName:
          type: string
        vetName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        isEmergency:
          type: boolean
        notes:
          type: string
          nullable: true